fix review comment.
