+++ src/main/java/jp/co/future/uroborosql/utils/StringFunction.java
@@ -511,16 +511,18 @@ public final class StringFunction {
 	 * 加算を行う. 最大値に対して加算した場合はオーバーフローする.<br>
 	 * 引数の型（short/int/longのいづれか）に合わせて計算を行う.
 	 *
+	 * @param <T> 引数の型
 	 * @param num 加算を行う数値
 	 * @return 加算後の数値（オーバーフローした場合は引数の型の最小値になる）
 	 */
-	public Object increment(final Object num) {
+	@SuppressWarnings("unchecked")
+	public <T extends Number> T increment(final T num) {
 		if (num instanceof Short) {
-			return incrementShort((Short) num);
+			return (T) incrementShort((Short) num);
 		} else if (num instanceof Integer) {
-			return incrementInt((Integer) num);
+			return (T) incrementInt((Integer) num);
 		} else if (num instanceof Long) {
-			return incrementLong((Long) num);
+			return (T) incrementLong((Long) num);
 		} else {
 			throw new IllegalArgumentException("Invalid value type. num must be of type short/int/long.");
 		}
@@ -532,7 +534,7 @@ public final class StringFunction {
 	 * @param num 加算を行う数値
 	 * @return 加算後の数値（オーバーフローした場合はshort型の最小値になる）
 	 */
-	public short incrementShort(final short num) {
+	public Short incrementShort(final Short num) {
 		return (short) (num + 1);
 	}
 
@@ -542,7 +544,7 @@ public final class StringFunction {
 	 * @param num 加算を行う数値
 	 * @return 加算後の数値（オーバーフローした場合はint型の最小値になる）
 	 */
-	public int incrementInt(final int num) {
+	public Integer incrementInt(final Integer num) {
 		return num + 1;
 	}
 
@@ -552,7 +554,7 @@ public final class StringFunction {
 	 * @param num 加算を行う数値
 	 * @return 加算後の数値（オーバーフローした場合はlong型の最小値になる）
 	 */
-	public long incrementLong(final long num) {
+	public Long incrementLong(final Long num) {
 		return num + 1L;
 	}
